changed file names and refactored to pass tests
