+++ packages/commonwealth/client/scripts/views/pages/view_proposal/index.tsx
@@ -33,8 +33,8 @@ import { CollapsibleProposalBody } from '../../components/collapsible_body_text'
 import { CWContentPage } from '../../components/component_kit/CWContentPage';
 import { VotingActions } from '../../components/proposals/voting_actions';
 import { VotingResults } from '../../components/proposals/voting_results';
+import { JSONDisplay } from './JSONDisplay';
 import { AaveViewProposalDetail } from './aave_summary';
-import { JSONDisplay } from './json_display';
 import type { SubheaderProposalType } from './proposal_components';
 import { ProposalSubheader } from './proposal_components';
